+++ docs/en_US/changelog.md
@@ -3,6 +3,9 @@
 ### TODO list
 - Add coloration to highlight data that are up to date in comparison with data that has not been updated since a while
 
+## 07/21/2020
+- Fixed a bug that kept the old top price, even if there was no value to display
+
 ## 07/20/2020
 - Add the oissubility to select the display order of favorites